powerproxy/requirements.txt
PyYAML==6.0.1
httpx==0.27.0
uvicorn[standard]==0.30.1
fastapi==0.111.0
tiktoken==0.7.0
azure-identity==1.17.1
azure-monitor-ingestion==1.0.4
redis[hiredis]==5.0.7
jsonschema==4.23.0
prometheus-fastapi-instrumentator~=6.0.0
